refactor(deleteUser): refacto hardDeleteUser to be able to use it as a standalone script. Add method to delete userAccountHash objects linked to a user.
