+++ back/src/users/database.ts
@@ -10,6 +10,7 @@ import { getUid, sanitizeEmail, hashToken } from "../utils";
 import { deleteCachedUserRoles } from "../common/redis/users";
 import { hashPassword, passwordVersion } from "./utils";
 import { UserInputError } from "../common/errors";
+import { PrismaTransaction } from "../common/repository/types";
 
 export async function getUserCompanies(userId: string): Promise<Company[]> {
   const companyAssociations = await prisma.companyAssociation.findMany({
@@ -54,6 +55,20 @@ export async function createUserAccountHash(
       role,
       companySiret: siret
     }
+  });
+}
+
+/**
+ * Delete UserAccountHash objects linked to a user's mail (association between user and siret with a hash)
+ * @param user
+ * @param prisma
+ */
+export async function deleteUserAccountHash(
+  user: User,
+  prisma: PrismaTransaction
+) {
+  await prisma.userAccountHash.deleteMany({
+    where: { email: user.email }
   });
 }
 
@@ -241,5 +256,126 @@ export async function updateUserPassword({
   return prisma.user.update({
     where: { id: userId },
     data: { password: hashedPassword, passwordVersion }
+  });
+}
+
+export async function checkCompanyAssociations(user: User): Promise<string[]> {
+  const errors: string[] = [];
+  const companyAssociations = await prisma.companyAssociation.findMany({
+    where: {
+      user: {
+        id: user.id
+      }
+    },
+    include: {
+      company: { select: { id: true, siret: true, vatNumber: true } }
+    }
+  });
+
+  for (const association of companyAssociations) {
+    if (association.role !== "ADMIN") {
+      continue;
+    }
+
+    const otherAdmins = await prisma.companyAssociation.findMany({
+      where: {
+        role: "ADMIN",
+        user: {
+          id: { not: user.id }
+        },
+        company: {
+          id: association.company.id
+        }
+      }
+    });
+    if (otherAdmins.length <= 0) {
+      errors.push(
+        `Impossible de supprimer cet utilisateur car il est le seul administrateur de l'entreprise ${
+          association.company.id
+        } (SIRET OU TVA: ${
+          association.company.siret ?? association.company.vatNumber
+        }).`
+      );
+    }
+  }
+
+  return errors;
+}
+
+export async function checkApplications(user: User): Promise<string[]> {
+  const errors: string[] = [];
+  const applications = await prisma.application.findMany({
+    where: {
+      adminId: user.id
+    }
+  });
+  for (const application of applications) {
+    errors.push(
+      `Impossible de supprimer cet utilisateur car il est le seul administrateur de l'application ${application.id} (${application.name}).`
+    );
+  }
+
+  return errors;
+}
+
+export async function deleteUserCompanyAssociations(
+  user: User,
+  prisma: PrismaTransaction
+) {
+  await prisma.companyAssociation.deleteMany({
+    where: {
+      user: {
+        id: user.id
+      }
+    }
+  });
+}
+
+export async function deleteMembershipRequest(
+  user: User,
+  prisma: PrismaTransaction
+) {
+  await prisma.membershipRequest.deleteMany({
+    where: {
+      user: {
+        id: user.id
+      }
+    }
+  });
+}
+
+export async function deleteUserActivationHashes(
+  user: User,
+  prisma: PrismaTransaction
+) {
+  await prisma.userActivationHash.deleteMany({
+    where: {
+      user: {
+        id: user.id
+      }
+    }
+  });
+}
+
+export async function deleteUserAccessTokens(
+  user: User,
+  prisma: PrismaTransaction
+) {
+  await prisma.accessToken.deleteMany({
+    where: {
+      user: {
+        id: user.id
+      }
+    }
+  });
+}
+
+export async function deleteUserGrants(user: User, prisma: PrismaTransaction) {
+  await prisma.grant.deleteMany({
+    where: {
+      user: {
+        id: user.id
+      }
+    }
   });
 }
